requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.8.4
chromadb==0.6.3
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.20.0
asyncio==3.4.3
gunicorn==21.2.0
PyPDF2==3.0.1
uvicorn==0.25.0
hypercorn==0.16.0 
